requirements.txt
whisper
torch
openai-whisper
soundfile
numpy
praat-parselmouth
python-Levenshtein
montreal-forced-aligner
nltk